requirements.txt
requests~=2.28.1
flask~=2.2.2
flask-sqlalchemy
flask_wtf
Flask-MySQL
pymysql
cryptography==38.0.3
python-dotenv~=0.21.0
WTForms~=3.0.1
pytest~=7.1.3
